starting to extract charm logic
moving to common-charm for now

allows llhl and jumble to share logic
