+++ typescript/packages/common-charm/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@commontools/lookslike-high-level",
+  "name": "@commontools/charm",
   "version": "0.1.0",
-  "description": "spa",
-  "main": "src/main.ts",
+  "description": "charm",
+  "main": "src/index.js",
   "type": "module",
   "packageManager": "pnpm@10.0.0",
   "engines": {
@@ -12,9 +12,7 @@
     "node": "20.11.0"
   },
   "scripts": {
-    "dev": "vite",
     "test": "vitest --exclude=lib/**/*.test.js",
-    "e2e": "playwright test",
     "format": "prettier --write . --ignore-path ../../../.prettierignore",
     "lint": "eslint .",
     "build": "tsc && vite build"
@@ -24,25 +22,14 @@
   "dependencies": {
     "@commontools/builder": "workspace:*",
     "@commontools/html": "workspace:*",
-    "@commontools/llm-client": "workspace:*",
-    "@commontools/os-ui": "workspace:*",
-    "@commontools/runner": "workspace:*",
-    "@commontools/iframe-sandbox": "workspace:*",
-    "@commontools/charm": "workspace:*",
-    "@commontools/ui": "workspace:*",
     "@commontools/memory": "workspace:*",
-    "lit": "^3.2.1",
-    "marked": "^15.0.6",
+    "@commontools/runner": "workspace:*",
     "merkle-reference": "^2.0.1",
-    "mermaid": "^11.4.1",
-    "panzoom": "^9.4.3",
-    "tslib": "^2.8.1",
     "zod": "^3.24.1",
     "zod-to-json-schema": "^3.24.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
-    "@playwright/test": "^1.50.0",
     "@types/node": "^22.10.10",
     "eslint": "^9.19.0",
     "eslint-config-prettier": "^10.0.1",
